+++ config-templates/module_perun_getMetadata.php
@@ -1,0 +1,31 @@
+<?php
+
+$config = array(
+	/**
+	 * Identifier of Proxy
+	 */
+	'proxyIdentifier' => '',
+
+	/**
+	 * Name of facility attribute Proxy Identifiers
+	 */
+	'perunProxyIdentifierAttr' => '',
+
+	/**
+	 * Name of facility attribute EntityID
+	 */
+	'perunProxyEntityIDAttr' => '',
+
+	/**
+	 * Absolute path, where the metadata will be stored
+	 */
+	'absoluteFileName' => '',
+
+	/**
+	 * List of attributes definitions
+	 */
+	'attributesDefinitions' => array(
+		// Name of attribute from perun => key which will be used in generated metadata
+		'perunAttrName' => 'metadataName',
+	),
+);